Update build targets and improve test execution
This commit reorders build targets in the Build.cs to make StagingPublishServices depend on CleanStagingDataBase and StagingCompile. Also, tests are updated for more robust execution: DispatcherPriorities are managed in the ObjectExtension and test cases in MainWindowTests and TextBoxExtension have been commented out or simplified. Other minor changes improve the clarity of the code.