Update Utf8JsonWriter and dependencies bindings
Updated the SetKestrel method in Utf8JsonWriterExtension.cs to take a 'domain' parameter and use it while setting paths for 'Certificate' and 'KeyPath'. Also, shifted the binding of IObjectStorage to SqliteObjectStorage from UiModule.cs to AndroidModule.cs and DesktopModule.cs. This change maintains segregation rules and also readies the system for future deployment or binding changes.