Add verification email feature with UI and tests
This commit introduces a verification email feature to the app. A new enum (UserIdentifierType) has been created for identification purposes, which can be either Email or Login. The interface IVerificationEmail, ViewModel VerificationCodeViewModel, and View VerificationCodeView are created pertaining to this. Changes also include necessary backend logic, UI updates with corresponding views and their rendering logic. Tests and utilities are updated to cover this feature.