Refactor EmailService and AuthenticationService
This commit primarily changes the email body parameter from 'subject' to 'text' in the EmailService for better clarity. It also integrates the sending of email in the AuthenticationService after verification code is generated. Configuration for email service removed from `appsettings.json` and `ServiceCollectionExtension.cs`. Some lines in 'ConfigurationManagerExtension.cs' is also consolidated for brevity.