Refactor VerificationCodeView layout
The layout of VerificationCodeView.axaml file was transformed in this commit. The changes include replacing the StackPanel with a controls:Card, and placing the buttons and TextBox inside a Grid for more organized structure. This is done to centralize and condense the verification code features for a cleaner and neater look.