Update email settings and verification code logic
Introduced test settings for email service and updated the email settings in appsettings.json. Refactored the email service related parts in the build scripts, removing MailServer and MailLogin parameters. Improved handling of email settings in JSON files to preserve existing settings not related to the password. Additionally, the VerificationCodeViewModel now handles calling the authentication service to update the verification code, a task which was previously done in the GrpcAuthenticationService. This commit also ensures that the test settings file is always copied to the output directory during the build.