Add task tracking and update test and build scripts
A static list of tasks has been added to the TaskWork model to keep track of all tasks. This list is updated every time a task is run. The changes include introducing an UpdateTasks method for this, and modifying the RunAsync methods to call it. Adjustments were made to the build script to include a new apk copy instruction. Test scripts have also been updated, with commented out lines indicating areas for future improvement.