Add SpravyVersion struct and refactor version handling
A new struct, SpravyVersion, has been created to encapsulate the handling and formatting of version numbers. All references to generating the version string have been replaced with the appropriate constructor or method from the new SpravyVersion struct. This refactor simplifies the version handling and makes the code easier to maintain.