Update theme selection mechanism in Account Settings
This change updates the way themes are selected in the Account Settings view. Instead of directly using the color theme, we've introduced a new model object to track the selected state of each color. This helps to highlight the currently active theme to the user. The view has also been adjusted to bind correctly to the new model state. We've also made minor adjustments to the UI layout in the App.axaml.