Refactor DayOfMonth and DayOfYear selectors in UI
The code has been refactored to enhance the DayOfMonth and DayOfYear selectors in the user interface. The selectors have been renamed to ToDoItemDayOfMonthSelector and ToDoItemDayOfYearSelector respectively. The changes include updates to view models and styles, and removal of unnecessary imports. Also, a new style for ToggleButton was added. The commit aims to provide a cleaner and more understandable code base.