Add migration to convert emails to upper case
The commit introduces a new migration to the authentication DB that converts all stored user email addresses to upper case. This standardizes email representation in database and prevents case sensitivity issues. Also, the changes include updating the `Login` and `Email` in `EfAuthenticationService.cs` to be trimmed and converted to upper case before being saved to the database. The product version in `SqliteSpravyAuthenticationDbContextModelSnapshot.cs` has also been updated to "8.0.2".