Simplify Compile method in Build.cs
The Compile method has been dramatically simplified to reduce complexity. The previous code executed a sequence of operations on projects which did not contain "Android" in their names and were not "Spravy.Ui.Desktop". Now these operations are simply referred to as 'project.Compile()', abstracting the detailed steps inside the 'Compile' method of the 'project' object. This change results in a much cleaner, easier to read and maintain method.