Update test execution process
The modifications in the code change the way tests are executed during the build process. Instead of executing a single test project, a foreach loop has been introduced to run tests on each test project. In order to facilitate this, an additional Test method has been added to the TestProjectBuilder class.