Add FTP functionality to DesktopProjectBuilder
A new class named DesktopProjectBuilderOptions was introduced to handle FTP options. A publish method with FTP functionality was added to the DesktopProjectBuilder class. The publish process was also updated in the Build.cs by changing the PublishDesktop method to use the new FTP options. Additionally, the DesktopProjectBuilderOptions instance is now created in the ProjectBuilderFactory.