Extend IPublished interface and update references
The IPublished interface has been extended to include ICsprojFile and IFtpOptions. In response to this change, all occurrences of 'PublishFolders' have been replaced with 'Publisheds' throughout the code in order to accommodate the new interface structure. The method for copying files in the BrowserProjectBuilder class has been updated to use 'published.GetAppFolder()'.