Move app folder content copy operation
The code modification involves a simple change in the sequence of operations performed in BrowserProjectBuilder. It specifically moves the command that copies the content from the application folder into the browser folder to be run right after creating the version folder. This rearrangement ensures these operations are performed in a logical and sequential manner.