Normalize user input in authentication service
Updated the user authentication service to normalize user input by making sure that all logins and emails are trimmed before processing. This change will ensure consistent and accurate processing of user information in the system regardless of how the data is input, thereby improving the reliability and integrity of the system. Furthermore, emails are also converted to upper case invariant for unique identification.