Simplify SettingsView layout structure
The code changes serve to simplify and streamline the SettingsView layout structure. The previous use of SettingsLayout and ObservableCollection has been removed and replaced with a more direct assembly of graphical components inside a GroupBox. This refactoring optimizes the item structure in the code and contributes to its readability and maintainability.