Add feature to clone To-Do items
A new functionality has been added to allow the cloning of To-Do items in the system. This includes updates in various services, UI components, and database transactions. Moreover, methods have been updated for better cancellation token handling. This feature improves the user's ability to duplicate tasks that they frequently create.