Refactor validation system and rename Error class to Result
The refactoring includes the relocation of validation related classes to a separate namespace and necessary changes in name and code to replace the usage of Error class with a new Result class. The Error class methods are also accordingly renamed. Also, the ErrorExtension from the Service is deleted and replaced with a newly created ResultExtension.