Refactor theme change method to run asynchronously
The theme change method in the SettingViewModel has been refactored to run in the background asynchronously. This will help improve user interface responsiveness by ensuring that the application doesn't hang while changing the theme. The method also sets a flag (`IsBusy`) to indicate if it is currently running, which can be used to prevent multiple concurrent executions.