Add feedback domain and update exception handling
A new feedback domain has been added that includes an interface for feedback and an error feedback model. In relation to this, the exception handling within the UI has also been updated. Notably, the TextBox of the ExceptionView now binds to the Message property, and the ExceptionViewModel has been amended to subscribe to the new Message property.