Update element references to use static constants
This update involves the modification of element reference naming in various .axaml and .cs files. Instead of directly using string names, static constants are now used to identify elements such as buttons, text boxes, and views. This change enhances maintainability and debuggability, as well as providing a centralized way of referring to these UI elements.