Add MaxCycleCountReachedError and update related methods
A new error class MaxCycleCountReachedError is created to raise an exception when maximum cycle count is reached. Changes are made to CycleHelper's While method to return a Result instead of void, enabling it to throw the new error. A method ThrowIfError is also added to the ResultExtension to facilitate error handling. Updates are reflected in ObjectExtension tests as well.