Refactor window.RunJobsAll usage in tests
The argument from window.RunJobsAll() method has been removed to simplify code in test files. This should not affect the functionality of the code as the specified argument was unnecessary and has been removed for cleaner code. The change has been applied to both 'VisualExtension.cs' and 'MainWindowTests.cs' files.