Update error handling and testing in UI and domain
The 'ValidationResults' property was renamed to 'Errors' in the 'ErrorViewModel' and its usage updated in the 'ErrorView' and 'ErrorHandler'. In 'WindowExtension', the method 'HasError' now accommodates for 'ErrorViewModel' along with 'ExceptionViewModel'. Simultaneously, some GUID values were updated in domain error classes. Furthermore, 'RunJobsAll' method in 'MainWindowTests' now runs four jobs instead of one.