Add SaveFrame calls in MainWindowTests
The calls to the SaveFrame method have been added at two different points in the MainWindowTests.cs file to handle the state of the application window for ongoing tests. With this addition, the state of the test window can be maintained even when an exception occurs or when the tests complete.