Refactor property setting and clean methods in project builders
Changed the 'AddProperty' method for 'Version' to use 'SetProperty', ensuring that the version is always updated in the AndroidProjectBuilder. Also moved the deletion of the Bin and Obj folders to occur after the 'DotNetClean' invocation in the Clean method of ProjectBuilder. Additionally, disabled the 'NoBuild' and 'NoRestore' methods in AndroidProjectBuilder to ensure they are always executed.