Refactor ToDo Items ViewModel and update commands
Removed "Add to Favorite" command from ToDo Items ViewModel and changed Id bindings to CurrentId in App.axaml. The refactor simplifies the command setup for each ToDo item in the ViewModel by eliminating the use of a separate favorite command. "Id" property in the ToDo items was replaced with "CurrentId" to provide more clarity about the current item being processed.