Updated handling of Reference ID in To-Do Items
This commit refactors the handling of Reference ID in todo item related methods and classes. Instead of directly assigning the Reference ID for every todo item, we now check if the item type is 'Reference' before proceeding. Furthermore, we've removed the direct assignment of Reference ID from several classes and methods where it wasn't necessary.