Update ToDo and PasswordItem libraries with async methods
The ToDo and PasswordItem libraries have been updated with asynchronous methods for a more efficient concurrent programming. This includes additions to the IToDoCache interface to get short items and update parent functionalities, as well as converting some methods in the ToDoCache, PasswordItemCache, and related classes to asynchronous tasks. Additionally, a small feature to populate Path array with default RootItem has been added to the ToDoItemViewModel.