Refactor To-Do cache to use TryAdd method
In this update, the method of adding elements to the `ToDoCache` service's `cache` and `activeCache` dictionaries was modified. Rather than directly adding items, the code now attempts to add them using the `TryAdd` method to avoid any potential errors or duplications. If the `TryAdd` operation doesn't succeed, it will return the existing item.