Update MemoryExtension and enhance ToDo feature
In the MemoryExtension, methods have been updated to primarily use Span instead of ReadOnlyMemory. Implementations of Join, Sum and Select were added and revised. In the LeafToDoItemsViewModel class, new injections were introduced and moved to a different part of the code. An update has also been made to enhance the ToDo feature to better check for empty selections and add multiple commands. Lastly, ToReadOnlyMemory helper method was added in the ObjectExtension class.