Add RandomizeChildrenOrder feature in ToDo list
A new feature has been added that provides the ability to randomize the order of children entity in ToDo list. This commit includes the updated UI views, models, and translation strings for the new functionality. In addition, minor adjustments have been made to some existing views for improved user experience.