Refactor button command and enhance view identification
Updated the button command in the ToDoItemView to use a more direct binding, removing unnecessary static command. Also, modified the UiApplicationService to include a 'GetCurrentViewType' function. This heavily aids in navigation by determining the current view type. An additional update was made to ToDoItemEntityNotify model to use this new function in determining actions, improving efficiency and flexibility.