Added IsMulti functionality to MultiToDoItemsView
This update enriches the MultiToDoItemsView by adding a new property, IsMulti, and correlating PseudoClasses functionality. Consequently, the UI can now react and adjust to changes in the multi-state of the ToDo items. A new StyledProperty is defined to facilitate two-way binding of this property, thus ensuring synchronization between the UI and the underlying ViewModel. This change also includes a modification in the ax