Update TextBox input method in integration tests
This commit refactors the way TextBox input is handled in integration tests. The "SetText" method in TextBoxExtension.cs has been modified to include an optional "focus" parameter, providing flexibility for scenarios in which the focus of the text box is not desired. Consequently, adjustments have been made in Tests.cs to accommodate these changes.