Add cleanup command and enhance folder management
Introduced a command to clean Docker images before the Docker build process. In addition, folder management in the BrowserProjectBuilder class has been enhanced by maintaining a separate 'currentFolder' for every project, in which the zip file of the project's application folder is stored. The same process is applied for every runtime environment as well.