Refactor SSH connection creation to use authentication methods
These changes aimed to refactor the SSH connection creation process to use a list of authentication methods instead of a single password. The modification was carried out in both the Build.cs and SshOptionsExtension.cs files. This not only cleans up the code but provides more flexibility for future additional authentication methods.