Refactor DayOfMonthSelector to use new integer selector controls
The DayOfMonthSelector feature in the UI has been refactored to use the new IntegerSelectorControl and IntegersSelectorControl. This removed redundancy in the code and allowed improved clarity and maintainability. Changes also included necessary style and template updates for these new controls. This refactoring improves the flexibility and modularity of the control usage.