Remove outdated mapper profiles and related classes
The commit deletes a set of outdated mapper profiles, interfaces and models which were no longer needed in the authentication, domain, event bus, password generator, schedule, and ToDo sections of the application. This clean up allows to reduce redundancy and maintenance effort by getting rid of unused code. Additional refactoring or repurposing of these codes might be planned for the future.