Implement new ObjectExtensionV2 for struct option
The commit refactors the ToOption extension method in different areas of the code to use the newly implemented ObjectExtensionV2, which is designed to handle struct objects. This simplifies the process of converting an Id to an option by removing the previous cast check. The code has been adjusted in SpravyCommand.cs, ToDoItemEntityNotify.cs, and EfToDoService.cs to use this new extension method.