Added "mode-press" class to buttons across the application
Literally, this enhances the look and feel of buttons across the application by adding a "mode-press" class to them. This change results in improved button visibility, adding a more interactive and aesthetic touch to the user interface. It involves different parts of the application, including settings, task management, password resetting, account management, and many more.