Add UiHelper and update SpravyCommandService
This commit includes the addition of a new UiHelper in the Helpers directory and significant updates to the SpravyCommandService. New command services have been added in the expanded file, including navigation, completion, favorites handling, linking, and more. Most of these commands also have a multi-item variant, effectively extending the functional range of the application.