Refactor SpravyCommand to simplify view handling
Removed the redundant view handling portion of the SpravyCommand methods in SpravyCommandService.cs. The view is now passed directly as a parameter which simplifies the code significantly. Modified the binding context in various views to bind the MainViewModel from the view rather than the DataContext. Removed MultiClone from the SpravyCommandNotifyService due to irrelevance.