Add theme initialization to Settings view
The commit introduces the theme initialization process on component load within the Settings view. It has refactored the theme-related logic from the SettingViewModel constructor to a new `SettingViewInitialized` command in SpravyCommandService, which is now invoked when the Settings view is initialized. A new command setter `SettingViewInitialized` was also added in the UiHelper class. A singleton instance of SukiTheme is now provided through IUiModule.