Refactor code to use GetSelectedItems method
The code in SpravyCommandService.cs has been refactored to replace the manual retrieval and checking of selected items with the use of the GetSelectedItems method. This change simplified the logic and removed the repeated code that was previously used in multiple places. As a result, the code is now more maintainable and easier to understand.