Refactor add ToDo item function in SpravyCommandService
The process of adding ToDo items in the SpravyCommandService has been refactored to improve its overall functionality. The code changes allow the application to handle multiple ToDo items simultaneously with the use of the 'IfSuccessForEachAsync' function. The ViewModel has been adjusted accordingly to handle the transformations associated with the refactored function.