Integrate ToDoService and ToDoCache into SpravyCommandService
The SpravyCommandService functions now include ToDoService and ToDoCache. The code has also been refactored to simplify the creation of LoginViewModel and to increase the use of serviceFactory within ToDoCache. As a result, the system will now be able to incorporate and handle ToDo items more efficiently, providing a better user experience.